refactor: more renaming and moving
