+++ src/cli/changesets/index.ts
@@ -1,12 +1,36 @@
+import { readPreState } from '@changesets/pre';
+import readChangesets from '@changesets/read';
+import type { NewChangeset, PreState } from '@changesets/types';
+
 import * as Git from '../../api/git';
 import { hasDebugFlag } from '../../utils/args';
 import { createLogger } from '../../utils/logging';
 
 import { runPublish } from './publish';
-import { readChangesetState } from './state';
 import { runVersion } from './version';
 
 const RELEASE_BRANCH_PREFIX = 'changeset-release/';
+export interface ChangesetState {
+  preState: PreState | undefined;
+  changesets: NewChangeset[];
+}
+
+const readChangesetState = async (dir: string): Promise<ChangesetState> => {
+  const preState = await readPreState(dir);
+  const isInPreMode = preState !== undefined && preState.mode === 'pre';
+
+  let changesets = await readChangesets(dir);
+
+  if (isInPreMode) {
+    const changesetsToFilter = new Set(preState.changesets);
+    changesets = changesets.filter((x) => !changesetsToFilter.has(x.id));
+  }
+
+  return {
+    preState: isInPreMode ? preState : undefined,
+    changesets,
+  };
+};
 
 export const changesetsRelease = async (args = process.argv) => {
   const cwd = process.cwd();
